+++ core/src/Streamly/Internal/Data/Stream/Bottom.hs
@@ -421,6 +421,7 @@ concatMapM f m = fromStreamD $ D.concatMapM (fmap toStreamD . f) (toStreamD m)
 -- >>> concatMap f = Stream.concatMapM (return . f)
 -- >>> concatMap f = Stream.concatMapWith Stream.append f
 -- >>> concatMap f = Stream.concat . fmap f
+-- >>> concatMap f = Stream.unfoldMany (Unfold.lmap f Unfold.fromStream)
 --
 {-# INLINE concatMap #-}
 concatMap ::Monad m => (a -> Stream m b) -> Stream m a -> Stream m b